Develop

See merge request lobnews/ComponentwarePraktikum2!30

+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/beans/repositories/ActionRepositoryBean.java
@@ -2,7 +2,7 @@ package de.fh_dortmund.inf.cw.surstwalat.dispatcher.beans.repositories;
 
 import javax.ejb.Stateless;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.Action;
+import de.fh_dortmund.inf.cw.surstwalat.common.model.Action;
 import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.ActionRepositoryLocal;
 
 /**
@@ -10,7 +10,7 @@ import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.ActionRepositoryLo
  * @author Johannes Heiderich
  */
 @Stateless
-public class ActionRepositoryBean extends DispatcherRepository<Action, Long> implements ActionRepositoryLocal {
+public class ActionRepositoryBean extends DispatcherRepository<Action, Integer> implements ActionRepositoryLocal {
 
     /**
      * Default constructor.
