+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/beans/repositories/PlayerRepositoryBean.java
@@ -1,15 +1,16 @@
 package de.fh_dortmund.inf.cw.surstwalat.dispatcher.beans.repositories;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.Player;
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.PlayerRepositoryLocal;
 import javax.ejb.Stateless;
+
+import de.fh_dortmund.inf.cw.surstwalat.common.model.Player;
+import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.PlayerRepositoryLocal;
 
 /**
  * Session Bean implementation class PlayerRepositoryBean
  * @author Johannes Heiderich
  */
 @Stateless
-public class PlayerRepositoryBean extends DispatcherRepository<Player, Long> implements PlayerRepositoryLocal {
+public class PlayerRepositoryBean extends DispatcherRepository<Player, Integer> implements PlayerRepositoryLocal {
 
     /**
      * Default constructor.
